Externalize run method in their own class
