+++ bin/configfile-scripts.js
@@ -5,13 +5,14 @@ const fs = require('fs')
 const path = require('path')
 
 const { runCommand, scriptsCommand } = require('../src/commands')
-const { ConfigService, FileService } = require("../src/services")
+const { ExecService, FileService, ConfigService } = require('../src/services')
 
 const packageData = JSON.parse(fs.readFileSync(path.join(__dirname, '../package.json')))
 const optionsFilePath = packageData.config.optionsFilePath.replace('~', process.env.HOME)
 
 const configService = new ConfigService(optionsFilePath)
 const fileService = new FileService(configService)
+const execService = new ExecService()
 
 program
   .version(packageData.version)
@@ -27,6 +28,6 @@ program
   .command('run <name>')
   .alias('r')
   .description('run custom configuration scripts.')
-  .action(runCommand(configService, fileService))
+  .action(runCommand(execService, fileService))
 
 program.parse(process.argv)
