+++ src/services/index.js
@@ -1,9 +1,11 @@
-const ateration = require('./ateration-logger')
+const alteration = require('./ateration-logger')
 const config = require('./config')
 const file = require('./file')
+const exec = require('./exec.service')
 
 module.exports = exports = {
   ...alteration,
   ...config,
-  ...file
+  ...file,
+  ...exec
 }
